Add more keys to remap
I just discovered this mode.  I had similar settings in my emacs init.  I would like to also change the keys for the up & down arrows to follow the up and down behavior, and the left and right keys to cycle through the history or do the ido-next-match-dir.  It basically switches the up and down and left and right keys.